requirements.txt
watchdog==3.0.0
requests==2.31.0
typing-extensions==4.9.0
google-api-python-client==2.120.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.1
